feat(utils.go): add ParseBody function to parse request body into a given struct
The `ParseBody` function is added to the `utils` package. This function takes an `http.Request` and a pointer to a struct as parameters. It reads the request body using `io.ReadAll` and then uses `json.Unmarshal` to parse the body into the provided struct. If there are any errors during the parsing, the function returns without modifying the struct.